requirements-linux.txt
# Requirements for Linux (Ubuntu) with PyMETIS support
# Install with: pip install -r requirements-linux.txt

# Core ML
torch>=2.0.0
torch-geometric>=2.3.0

# Graph partitioning (Linux only - requires METIS library)
# Install METIS first: sudo apt-get install libmetis-dev
pymetis>=2023.1

# Other dependencies
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
pandas>=2.0.0
matplotlib>=3.7.0

# Post-Quantum Cryptography
kyber-py>=0.1.0
pycryptodome>=3.18.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Testing
pytest>=7.3.0
